Converted EiffelGraph to full Void-safety (and also the egraph example).
One minor breaking change is the removal of `EG_VECTOR2D.default_create`, now always use `make (.., ..)`

git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@99702 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/examples/vision2/egraph/main_window.e
@@ -174,7 +174,6 @@ feature {NONE} -- Implementation, Close event
 			-- The user wants to close the window
 		local
 			question_dialog: EV_CONFIRMATION_DIALOG
-			l_app: detachable EV_APPLICATION
 		do
 			create question_dialog.make_with_text (Label_confirm_close_window)
 			question_dialog.show_modal_to_window (Current)
@@ -184,9 +183,11 @@ feature {NONE} -- Implementation, Close event
 				destroy;
 
 					-- End the application
-				l_app := (create {EV_ENVIRONMENT}).application
-				check l_app /= Void end -- Implied by application is running
-				l_app.destroy
+				if attached (create {EV_ENVIRONMENT}).application as l_app then
+					l_app.destroy
+				else
+					check application_is_running: False end
+				end
 			end
 		end
